Final Game
I commented my code, tweaked a few numbers and fixed some bugs.  The
game is not perfect, but I think it is as good as I can make it.  I
looked at the feedback.  I decided to leave the laser as it was, as I
couldn’t buff it without it becoming overpowered.  I made the powerup
light up when activated.  The ball still breaks it to prevent bugs.